fix: widen MSSP UPTIME narrowing cast to survive year 2038
GetMSSP_Uptime formatted s_Uptime (a time_t boot-epoch timestamp) through
sprintf(Buffer, "%d", (int)s_Uptime). Per the MSSP spec, UPTIME correctly
means the server's boot-time epoch (not elapsed uptime) -- s_Uptime being
set once from time(0) at startup and never updated is correct as-is; the
only defect was the (int) narrowing. Widen to static_cast<long long> + %lld
so the value survives once epoch seconds exceed INT_MAX. No observable
output change until the year 2038.

Not pinned with a dedicated test: GetMSSP_Uptime is a static (internal-
linkage) helper only reachable through SendMSSP, which writes straight to
a real socket descriptor via Write() -- not reachable from unit tests
without significant socket-mocking scaffolding, which is disproportionate
to a mechanical width fix.

+++ src/protocol.cpp
@@ -2383,7 +2383,13 @@ static const char* GetMSSP_Players()
 static const char* GetMSSP_Uptime()
 {
     static char Buffer[32];
-    sprintf(Buffer, "%d", (int)s_Uptime);
+    // UPTIME per the MSSP spec is the server's boot-time epoch (seconds since the Unix
+    // epoch when the server started), NOT elapsed uptime -- s_Uptime is set once from
+    // time(0) at startup and never updated, which is correct as-is. The only defect
+    // here was the narrowing (int) cast: widen to long long / %lld so the value doesn't
+    // wrap once epoch seconds exceed INT_MAX (the year-2038 problem). No observable
+    // output change until then.
+    sprintf(Buffer, "%lld", static_cast<long long>(s_Uptime));
     return Buffer;
 }
